frontend/requirements.txt
tqdm
numpy
pillow
requests
streamlit
opencv-python
ffmpeg-python
python-multipart